"修复 DateTime 纳秒检查整数溢出问题，使用 static_cast 确保大数计算安全"

+++ include/base/time/DateTime.cpp
@@ -54,7 +54,7 @@ void base::DateTime::CheckSecond()
 
 void base::DateTime::CheckNanosecond()
 {
-	if (_nanosecond < 0 || _nanosecond >= 1000 * 1000 * 1000)
+	if (_nanosecond < 0 || _nanosecond >= static_cast<int64_t>(1000) * 1000 * 1000)
 	{
 		throw std::invalid_argument{CODE_POS_STR + "非法纳秒。"};
 	}
